requirements.txt
pandas
beautifulsoup4
lxml
streamlit
leia-br